renames CollectionsRails -> CollectionHubRails
